JBIDE-21301 Preferences: path to OC binary is not shown on OpenShift 3 page if binary is on $PATH
